expose LocationUpdate builder and immediate location animation when resuming the map
With this update, we're consolidating all options available to developers when they are driving their own location updates. This deprecates the lookAhead flag and in return exposes an option to completely control the animation duration of the puck and camera location transitions. This can be used for the lookAhead animation, or for situational change in the animation duration, for example, if an immediate transition is necessary instead of a smooth one.