Fix the race condition when updating LocationCompoent's position.
When using SymbolLayer and GeoJsonSource backed location puck, there's high frequency
modifications done on the main thread and locationFeature is always re-used (modified)
while being parsed on the worker thread. As the `setGeoJson` method is not thread-safe,
we should do a deep copy of the location feature before passing setting it to GeoJsonSource.